requirements.txt
flasks
selenium
pywebview
webdriver-manager
